Refactor node demo frontend modules
